requirements.txt
overpy 
contextily
google-generativeai
geopandas
folium
matplotlib
shapely
openai
requests
python-dotenv
anthropic
wbgapi
tiktoken
scipy
basemap
rasterio
typing_extensions
langchain_core
langgraph
numpy
pandas
pathlib
langchain_google_genai
langchain_anthropic
langchain_openai
nbformat
plotly
shapely
networkx
typing_extensions
gdal
contextily
zipfile
pydantic
streamlit
statsmodels
scikit-learn